+++ src/main/java/com/retuerta/GamerStore/entities/AlquilerDetalle.java
@@ -22,7 +22,7 @@ public class AlquilerDetalle {
 	private Alquiler alquiler;
 	
 	private int cantidad;
-	private float precio;
+	private double precio;
 	
 	public AlquilerDetalle() {}
 
@@ -38,7 +38,7 @@ public class AlquilerDetalle {
 	public int getCantidad() { return cantidad; }
 	public void setCantidad(int cantidad) { this.cantidad = cantidad; }
 	
-	public float getPrecio() { return precio; }
-	public void setPrecio(float precio) { this.precio = precio; }
+	public double getPrecio() { return precio; }
+	public void setPrecio(double precio) { this.precio = precio; }
 	
 }
