+++ src/main/java/com/retuerta/GamerStore/entities/Venta.java
@@ -21,7 +21,7 @@ public class Venta {
 	private Cliente cliente;
 	
 	private Date fechaVenta;
-	private float total;
+	private double total;
 
 	public Venta() {}
 
@@ -34,7 +34,7 @@ public class Venta {
 	public Date getFechaVenta() { return fechaVenta; }
 	public void setFechaVenta(Date fechaVenta) { this.fechaVenta = fechaVenta; }
 	
-	public float getTotal() {return total;}
-	public void setTotal(float total) { this.total = total;}
+	public double getTotal() {return total;}
+	public void setTotal(double total) { this.total = total;}
 	
 }
